% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{components.fbl_bsts}
\alias{components.fbl_bsts}
\title{Extract meaningful components}
\usage{
\method{components}{fbl_bsts}(object, ...)
}
\arguments{
\item{object}{An estimated model.}

\item{...}{Unused.}
}
\value{
A \code{\link[fabletools:dable]{fabletools::dable()}} containing estimated states.
}
\description{
A bsts model consists of terms which are additively included in the model.
}
\details{
Extracting a bsts model's components using this function allows you to
visualise the components in a similar way to \code{\link[bsts:plotBstsComponents]{bsts::plotBstsComponents()}}.
}
\examples{

if (requireNamespace("tsibbledata")) {
library(tsibble)
library(dplyr)
beer_components <- tsibbledata::aus_production \%>\%
  model(
    bsts = BSTS(Beer ~ season("year"))
  ) \%>\%
  components()

beer_components

\dontrun{
autoplot(beer_components)

library(ggplot2)
library(lubridate)
beer_components \%>\%
  ggplot(aes(x = quarter(Quarter), y = year, group = year(Quarter))) +
  geom_line()
}
}

}
