% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{package}
\name{fable.bsts-package}
\alias{fable.bsts}
\alias{fable.bsts-package}
\title{fable.bsts: Bayesian Structural Time Series Modelling Interface for 'fable'}
\description{
Allows bsts models from the 'bsts' package to be used in a tidy workflow with the modelling interface of 'fabletools'. This extends 'bsts' to provide enhanced model specification and management, performance evaluation methods, and model combination tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davidtedfordholt/fable.bsts}
  \item Report bugs at \url{https://github.com/davidtedfordholt/fable.bsts/issues}
}

}
\author{
\strong{Maintainer}: David Holt \email{d@tdfd.org}

Other contributors:
\itemize{
  \item Steven Scott (bsts library, https://sites.google.com/view/stevethebayesian/software) [contributor]
}

}
\keyword{package}
