% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fitted.fbl_bsts}
\alias{fitted.fbl_bsts}
\title{Extract fitted values}
\usage{
\method{fitted}{fbl_bsts}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}
}
\value{
A vector of fitted values.
}
\description{
Extracts the fitted values from an estimated bsts model.
}
