% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{forecast.fbl_bsts}
\alias{forecast.fbl_bsts}
\title{Produce forecasts from the bsts model}
\usage{
\method{forecast}{fbl_bsts}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{new_data}{A \code{tsibble} containing future information used to forecast.}

\item{specials}{(passed by \code{\link[fabletools:forecast.mdl_df]{fabletools::forecast.mdl_df()}}).}

\item{...}{Additional arguments passed to \code{\link[bsts:predict.bsts]{bsts::predict.bsts()}}.}
}
\value{
A list of forecasts.
}
\description{
If additional future information is required (such as exogenous variables)
by the model, then they should be included as variables of the \code{new_data} argument.
}
\examples{

if (requireNamespace("tsibbledata")) {
library(tsibble)
library(dplyr)
tsibbledata::aus_production \%>\%
  model(
    bsts = BSTS(Beer ~ season("year"))
  ) \%>\%
  forecast()
}

}
\seealso{
\code{\link[bsts:predict.bsts]{bsts::predict.bsts()}}
}
