% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{glance.fbl_bsts}
\alias{glance.fbl_bsts}
\title{Glance a bsts model}
\usage{
\method{glance}{fbl_bsts}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\value{
A one row tibble summarising the model's fit.
}
\description{
A glance of a bsts provides the residual's standard deviation (sigma), and
a tibble containing the selected changepoints with their trend adjustments.
}
\examples{

if (requireNamespace("tsibbledata")) {
library(tsibble)
library(dplyr)
fit <- tsibbledata::aus_production \%>\%
  model(
    bsts = BSTS(Beer ~ season())
  )

glance(fit)
}

}
