% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_sum.fbl_bsts}
\alias{model_sum.fbl_bsts}
\title{Glance a bsts model}
\usage{
\method{model_sum}{fbl_bsts}(x)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}
}
\value{
A one row tibble summarising the model's fit.

A tibble containing the model's estimated parameters.
}
\description{
A glance of a bsts provides the residual's standard deviation (sigma), and
a tibble containing the selected changepoints with their trend adjustments.
}
\examples{

if (requireNamespace("tsibbledata")) {
library(tsibble)
library(dplyr)
fit <- tsibbledata::aus_production \%>\%
  model(
    bsts = BSTS(Beer ~ season())
  )

glance(fit)
}


if (requireNamespace("tsibbledata")) {
library(tsibble)
library(dplyr)
fit <- tsibbledata::aus_production \%>\%
  model(
    bsts = BSTS(Beer ~ season())
  )

tidy(fit) # coef(fit) or coefficients(fit) can also be used
}

}
