% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{residuals.fbl_bsts}
\alias{residuals.fbl_bsts}
\title{Extract model residuals}
\usage{
\method{residuals}{fbl_bsts}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}
}
\value{
A vector of residuals.
}
\description{
Extracts the residuals from an estimated bsts model.
}
